+++ CHANGELOG.md
@@ -11,6 +11,7 @@ LR(0) parser Change Log
 *   Add: Terminals can be defined with RegExp literal in grammar text.
 *   Add: Named fixed tokens in `VovanVE\parser\lexer\Lexer`.
 *   Add: Separate definition for inline tokens in `VovanVE\parser\lexer\Lexer`.
+*   Enh: RegExp validation throws exception from `VovanVE\parser\lexer\Lexer`.
 *   Deprecated: `VovanVE\parser\lexer\Lexer` constructor arguments should be avoided
     in favor to corresponding extending method.
 *   Deprecated: Method `VovanVE\parser\lexer\Lexer::extend()` - use specific corresponding